refactor(assembly): make multi-package directory assembly registry-based
Two ecosystem-specific directory splitters had accreted inside the generic
`sibling_merge.rs` engine as a hardcoded `if let Some(cocoapods)… if let
Some(maven)…` chain plus config-sniffing helpers. Replace the one-offs with the
two assembly injection points the architecture already implies:

- Generalize the Maven distinct-GAV split into `assemble_siblings_per_identity`
  and expose it as `AssemblyMode::SiblingMergePerIdentity`. Maven is now pure
  config (mode change in `assemblers.rs`); the Maven-specific function and
  `is_maven_sibling_config` gate are deleted. Any future directory that may hold
  several independent manifests opts in declaratively.
- Move the bespoke CocoaPods logic (primary-podspec selection + Podfile/lock
  attachment) into a dedicated `cocoapods_merge.rs`, registered for
  `CocoapodsPodspec` via `special_directory_merger_for` — the same registry that
  already routes Debian/Hackage/Hugging Face/Windows-update directories.

`sibling_merge.rs` keeps only the generic engine plus shared helpers (now
`pub(super)`); it carries no per-ecosystem branches. ADR 0006 documents the
directory-assembly-strategy taxonomy and the two injection points.

Pure refactor: assembly/maven/cocoapods lib tests and all 36 assembly goldens
(including maven-basic, maven-meta-inf-basic, and maven-distinct-gav-poms) pass
unchanged.
